requirements.txt
asgiref==3.5.0
black==23.3.0
certifi==2021.10.8
click==8.1.3
cloudinary==1.29.0
coverage==6.3.2
Django==4.0.3
django-crispy-forms==1.14.0
django-etc==1.4.0
django-hitcount==1.3.5
gunicorn==20.1.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.0.1
platformdirs==3.8.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2021.5
urllib3==1.26.9
whitenoise==6.0.0
